requirements.txt
Flask==0.10.1
Flask-Bootstrap==3.3.5.2
Flask-Docker==0.2.0
Flask-Redis==0.1.0
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.10.4
docker-py==1.2.3
itsdangerous==0.24
redis==2.10.3
requests==2.7.0
six==1.9.0
